Update methods for delete and confirm

+++ Lab4/mysite/profile_user/views.py
@@ -5,11 +5,18 @@ from hotel.models import Room, Booking
 
 # Create your views here.
 def get_book(request):
-    all_booking = Booking.objects.filter(client_id=request.user, status='cart')
-    return render(request, 'profile_user/profile.html', {'all_booking': all_booking})
+    def get_booking():
+        return Booking.objects.filter(client_id=request.user, status='cart')
 
+    if request.method == 'POST' and 'delete_item' in request.POST:
+        booking = get_object_or_404(Booking, id=request.POST['booking_id'])
+        booking.delete()
+        return render(request, 'profile_user/profile.html', {'all_booking': get_booking()})
 
-def delete_booking(request, booking_id):
-    booking = get_object_or_404(Booking, id=booking_id)
-    booking.delete()
-    return get_book(request)
+    if request.method == 'POST' and 'confirm' in request.POST:
+        for i in Booking.objects.filter(client_id=request.user, status='cart'):
+            i.status = 'paid'
+            i.save()
+        return render(request, 'profile_user/profile.html', {'all_booking': get_booking()})
+
+    return render(request, 'profile_user/profile.html', {'all_booking': get_booking()})
